+++ scripts/test/Ticket/TicketHistory.t
@@ -18,6 +18,12 @@ my $QueueObject  = $Kernel::OM->Get('Kernel::System::Queue');
 my $TypeObject   = $Kernel::OM->Get('Kernel::System::Type');
 my $StateObject  = $Kernel::OM->Get('Kernel::System::State');
 
+# Turn on the ticket type feature.
+$Kernel::OM->Get('Kernel::Config')->Set(
+    Key   => 'Ticket::Type',
+    Value => 1,
+);
+
 $Kernel::OM->Get('Kernel::System::Cache')->CleanUp();
 
 my @Tests = (
@@ -85,6 +91,19 @@ my @Tests = (
                         HistoryType => 'NewTicket',
                         Type        => 'Unclassified',
                     },
+
+                    # Bug 12702 - TicketHistoryGet() initial ticket type update
+                    {
+                        CreateBy    => 1,
+                        HistoryType => 'TypeUpdate',
+                        Queue       => 'Raw',
+                        OwnerID     => 1,
+                        PriorityID  => 3,
+                        State       => 'new',
+                        Type        => 'Unclassified',
+                        TypeID      => '1',
+                    },
+
                     {
                         CreateBy    => 1,
                         HistoryType => 'CustomerUpdate',
@@ -304,6 +323,7 @@ for my $Test (@Tests) {
 
             my %LookForHistoryTypes = (
                 NewTicket      => 1,
+                TypeUpdate     => 1,
                 OwnerUpdate    => 1,
                 CustomerUpdate => 1,
             );